# docs/openapi-object-detection.yaml
openapi: 3.0.3
info:
  title: FlowGuard Object Detection & SecurePi Integration API
  description: >
    JSON REST endpoints for Camera Inventory, Monitoring Zones (Detection Setup),
    Detection Alerts, and the SecurePi edge-ingest route. Derived from
    current server/routes/*.js and ai-service/main.py, audited 2026-07-28.
    MJPEG endpoints (SecurePi /video_feed, AI Service /api/yolo/stream) are
    intentionally excluded — they are not JSON REST responses.
  version: "1.0.0"
servers:
  - url: http://localhost:5001
    description: FlowGuard Node.js backend (dev)
...
components:
  securitySchemes:
    bearerJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: FlowGuard user JWT (FM/Staff/Tenant), verified against APP_SECRET.
    serviceKey:
      type: apiKey
      in: header
      name: x-service-key
      description: Shared secret used by the Python AI engine (AI_SERVICE_KEY).
    edgeIngestToken:
      type: http
      scheme: bearer
      description: Static device token used by the SecurePi edge node (EDGE_INGEST_TOKEN).
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      example:
        error: "zone_name and camera_location are required."
...
    PeopleCount:
      type: object
      properties:
        count: { type: integer }
        detection_active: { type: boolean }
        camera_status: { type: string }
    AnalyzeFrameInput:
      type: object
      required: [image]
      properties:
        image:
          type: string
          maxLength: 8388608
          pattern: '^data:image/(jpeg|png|webp);base64,'
          description: JPEG, PNG, or WebP base64 data URL captured from the browser camera or an uploaded video.
        camera_id:
          type: integer
          minimum: 1
          description: Optional Camera database identifier.
        zone_id:
          type: integer
          minimum: 1
          description: Optional MonitoringZone database identifier.
        source:
          type: string
          description: Optional user-visible capture source label.
    AnalyzeFrameResult:
      type: object
      properties:
        count: { type: integer }
        detections:
          type: array
          items:
            type: object
            properties:
              label: { type: string }
              box:
                type: array
                items: { type: integer }
                minItems: 4
                maxItems: 4
              status: { type: string }
              confidence: { type: number }
              type: { type: string }
        frame_width: { type: integer }
        frame_height: { type: integer }
        detection_active: { type: boolean }
        camera_status: { type: string }
...
paths:
...
  /api/yolo/people-count:
    get:
      tags: [AIService]
      summary: Poll current people count from the AI service's YOLO loop
      description: >
        Node JWT/RBAC proxy to the private FastAPI people-count endpoint. Node
        adds the service credentials; the browser never calls FastAPI directly.
      security: [{ bearerJwt: [] }]
      responses:
        '200':
          description: Current people count and detection state
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PeopleCount' }
        '401': { description: Missing JWT }
        '403': { description: Invalid JWT or role is not FM/Staff }
        '429': { description: AI proxy rate limit exceeded }
        '502': { description: AI service returned an error }
        '503': { description: AI service unavailable }
  /api/yolo/analyze-frame:
    post:
      tags: [AIService]
      summary: Analyze one browser-camera or uploaded-video frame
      description: >
        Node JWT/RBAC proxy to the private FastAPI frame-analysis endpoint.
      security: [{ bearerJwt: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AnalyzeFrameInput' }
      responses:
        '200':
          description: Detections for the analyzed frame
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AnalyzeFrameResult' }
        '400':
          description: Missing/invalid data-URL image or AI rejected the image
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401': { description: Missing JWT }
        '403': { description: Invalid JWT or role is not FM/Staff }
        '413': { description: Image payload too large }
        '429': { description: AI proxy rate limit exceeded }
        '502': { description: AI service returned an error }
        '503': { description: AI service unavailable }